options/posix: Move strcasestr() to posix
