+++ options/ansi/generic/string-stubs.cpp
@@ -1,5 +1,4 @@
 #include <string.h>
-#include <strings.h>
 #include <errno.h>
 #include <wchar.h>
 #include <ctype.h>
@@ -222,18 +221,6 @@ char *strchrnul(const char *s, int c) {
 		i++;
 	}
 	return const_cast<char *>(s + i);
-}
-
-char *strcasestr(const char *s, const char *pattern) {
-	size_t plen = strlen(pattern);
-	const char *p = s;
-	while(*p) {
-		// Need strncasecmp() to avoid checking past the end of a successful match.
-		if(!strncasecmp(p, pattern, plen))
-			return const_cast<char *>(p);
-		++p;
-	}
-	return nullptr;
 }
 
 double wcstod(const wchar_t *__restrict, wchar_t **__restrict) MLIBC_STUB_BODY
